+++ internal/versions/versions.go
@@ -11,7 +11,6 @@ import (
 	"strings"
 
 	"github.com/asdf-vm/asdf/internal/config"
-	"github.com/asdf-vm/asdf/internal/execenv"
 	"github.com/asdf-vm/asdf/internal/hook"
 	"github.com/asdf-vm/asdf/internal/installs"
 	"github.com/asdf-vm/asdf/internal/plugins"
@@ -24,7 +23,8 @@ const (
 	systemVersion           = "system"
 	latestVersion           = "latest"
 	uninstallableVersionMsg = "uninstallable version: %s"
-	latestFilterRegex       = "(?i)(^Available versions:|-src|-dev|-latest|-stm|[-\\.]rc|-milestone|-alpha|-beta|[-\\.]pre|-next|(a|b|c)[0-9]+|snapshot|master)"
+	latestFilterRegex       = "(?i)(^Available versions:|-src|-dev|-latest|-stm|[-\\.]rc|-milestone|-alpha|-beta|[-\\.]pre|-next|(a|b|c)[0-9]+|snapshot|master|main)"
+	numericStartFilterRegex = "^\\s*[0-9]"
 	noLatestVersionErrMsg   = "no latest version found"
 )
 
@@ -49,6 +49,17 @@ func (e NoVersionSetError) Error() string {
 	// with this improvement
 	// return fmt.Sprintf("no version set for plugin %s", e.toolName)
 	return "no version set"
+}
+
+// VersionAlreadyInstalledError is returned whenever a version is already
+// installed.
+type VersionAlreadyInstalledError struct {
+	toolName string
+	version  toolversions.Version
+}
+
+func (e VersionAlreadyInstalledError) Error() string {
+	return fmt.Sprintf("version %s of %s is already installed", e.version.Value, e.toolName)
 }
 
 // InstallAll installs all specified versions of every tool for the current
@@ -94,13 +105,16 @@ func Install(conf config.Config, plugin plugins.Plugin, dir string, stdOut io.Wr
 	}
 
 	for _, version := range versions.Versions {
-		err := InstallOneVersion(conf, plugin, version, false, stdOut, stdErr)
-		if err != nil {
-			return err
+		iErr := InstallOneVersion(conf, plugin, version, false, stdOut, stdErr)
+		var vaiErr VersionAlreadyInstalledError
+		if errors.As(iErr, &vaiErr) {
+			err = errors.Join(err, iErr)
+		} else if iErr != nil {
+			return iErr
 		}
 	}
 
-	return nil
+	return err
 }
 
 // InstallVersion installs a version of a specific tool, the version may be an
@@ -143,18 +157,17 @@ func InstallOneVersion(conf config.Config, plugin plugins.Plugin, versionStr str
 	installDir := installs.InstallPath(conf, plugin, version)
 
 	if installs.IsInstalled(conf, plugin, version) {
-		return fmt.Errorf("version %s of %s is already installed", version, plugin.Name)
+		return VersionAlreadyInstalledError{version: version, toolName: plugin.Name}
 	}
 
+	concurrency, _ := conf.Concurrency()
 	env := map[string]string{
 		"ASDF_INSTALL_TYPE":    version.Type,
 		"ASDF_INSTALL_VERSION": version.Value,
 		"ASDF_INSTALL_PATH":    installDir,
 		"ASDF_DOWNLOAD_PATH":   downloadDir,
-		"ASDF_CONCURRENCY":     asdfConcurrency(conf),
+		"ASDF_CONCURRENCY":     concurrency,
 	}
-
-	env = execenv.MergeEnv(execenv.SliceToMap(os.Environ()), env)
 
 	err = os.MkdirAll(downloadDir, 0o777)
 	if err != nil {
@@ -215,21 +228,6 @@ func InstallOneVersion(conf config.Config, plugin plugins.Plugin, versionStr str
 	return nil
 }
 
-func asdfConcurrency(conf config.Config) string {
-	val, ok := os.LookupEnv("ASDF_CONCURRENCY")
-
-	if !ok {
-		val, err := conf.Concurrency()
-		if err != nil {
-			return "1"
-		}
-
-		return val
-	}
-
-	return val
-}
-
 // Latest invokes the plugin's latest-stable callback if it exists and returns
 // the version it returns. If the callback is missing it invokes the list-all
 // callback and returns the last version matching the query, if a query is
@@ -249,7 +247,8 @@ func Latest(plugin plugins.Plugin, query string) (version string, err error) {
 			return version, err
 		}
 
-		versions := filterOutByRegex(allVersions, latestFilterRegex)
+		versions := filterOutByRegex(allVersions, numericStartFilterRegex, true)
+		versions = filterOutByRegex(versions, latestFilterRegex, false)
 
 		if len(versions) < 1 {
 			return version, errors.New(noLatestVersionErrMsg)
@@ -260,7 +259,8 @@ func Latest(plugin plugins.Plugin, query string) (version string, err error) {
 
 	// parse stdOut and return version
 	allVersions := parseVersions(stdOut.String())
-	versions := filterOutByRegex(allVersions, latestFilterRegex)
+	versions := filterOutByRegex(allVersions, numericStartFilterRegex, true)
+	versions = filterOutByRegex(versions, latestFilterRegex, false)
 	if len(versions) < 1 {
 		return version, errors.New(noLatestVersionErrMsg)
 	}
@@ -348,10 +348,11 @@ func filterByExactMatch(allVersions []string, pattern string) (versions []string
 	return versions
 }
 
-func filterOutByRegex(allVersions []string, pattern string) (versions []string) {
+func filterOutByRegex(allVersions []string, pattern string, keepMatch bool) (versions []string) {
+	regex, _ := regexp.Compile(pattern)
 	for _, version := range allVersions {
-		match, _ := regexp.MatchString(pattern, version)
-		if !match {
+		match := regex.MatchString(version)
+		if match == keepMatch {
 			versions = append(versions, version)
 		}
 	}